ril: use `shouldIgnoreRequest` for debuging
In some cases, the rild would crash when we send it a unsupported request.
We could ignore that request by override the method that send that request
and response a unsupported message to the caller. Doing just that for just
some methods is just fine. But in the case we want to debug what method cause
rild crash, there are more methods to override. This `shouldIgnoreRequest`
provides what we want. When we want to ignore any request, we just need to
set a system prop, no more recompiling the whole code base, which is time a
consuming progress.

Credit: @pawitp for inspiring me the idea about using system prop

Change-Id: Ia896bf45e76ae4adc0004a214313870abf19d9d6