recovery: do not use twrp.fstab as recovery.fstab
TWRP uses old fstab format which the current build system does not understand.
So TWRP uses a workaround to solve this issue that is it uses a recovery
fstab named twrp.fstab instead of recovery.fstab. Then, when we boot into
recovery mode, TWRP will rename twrp.fstab to recovery.fstab, and uses that
file as recovery fstab.

Change-Id: I02e33b8533842abcc0422bc5fcee572fd5c1844d